Change ValueType from int to double

+++ VectorNew.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <iostream>
-using ValueType = int;
+using ValueType = double;
 //! \brief Класс, реализующий С++ контейнер vector для типа int
 class Vector
 {
